Update approach to extracting SAMLResponse
After authentication, Okta will create an interstitial page with a form
including a hidden input with the SAML response data. Javascript then
immediately submits the form once the page loads. This means there is
limited time to extract the SAML response. Each `wait` call is capable
of detecting the SAML response. However, then executing an `evaluate`
call takes long enough for the SAML response to no longer be available.
So, we need to capture the SAML response in the `wait` call. The easiest
mechanism to get data from the browser back to Node is via the console
logger.

Rather than using wait/evaluate this method uses the wait to capture the
SAMLResponse and then the console log to pass it back to Node.